resolved iteration comments
