+++ Microsoft.Azure.Cosmos/src/ChangeFeed/ChangeFeedStartFromVisitor{TResult}.cs
@@ -1,0 +1,15 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.ChangeFeed
+{
+    internal abstract class ChangeFeedStartFromVisitor<TResult>
+    {
+        public abstract TResult Visit(ChangeFeedStartFromNow startFromNow);
+        public abstract TResult Visit(ChangeFeedStartFromTime startFromTime);
+        public abstract TResult Visit(ChangeFeedStartFromContinuation startFromContinuation);
+        public abstract TResult Visit(ChangeFeedStartFromBeginning startFromBeginning);
+        public abstract TResult Visit(ChangeFeedStartFromContinuationAndFeedRange startFromContinuationAndFeedRange);
+    }
+}
